fix: restructure code and show content based on permission (#261)
* fix: restructure code and show content based on permission
- Restructure Invite user forms into separate App and Org forms
- Show Modal content based on permissions.

* update: Permission handling.

* fix: Modify permission handling

* fix: check manage permission for manage users btn and show copy to clipboard always.

* Dummy commit to run the tests

* Fix: Test cases

* Another dummy commit to run the tests
