+++ app/client/cypress/integration/Smoke_TestSuite/OrganisationTests/CreateOrgTests_spec.js
@@ -58,7 +58,7 @@ describe("Create new org and share with a user", function() {
     cy.get(homePage.searchInput).type(appid);
     cy.wait(2000);
     cy.contains(orgid);
-    cy.xpath(homePage.ShareBtn).should("not.be.visible");
+    cy.xpath(homePage.ShareBtn).should("be.visible");
     cy.get(homePage.appEditIcon)
       .first()
       .click({ force: true });
