+++ app/client/src/pages/Applications/permissionHelpers.tsx
@@ -4,6 +4,8 @@ export enum PERMISSION_TYPE {
   MANAGE_APPLICATION = "manage:applications",
   READ_APPLICATION = "read:applications",
   READ_ORGANIZATION = "read:organizations",
+  INVITE_USER_TO_ORGANIZATION = "inviteUsers:organization",
+  MAKE_PUBLIC_APPLICATION = "makePublic:applications",
 }
 
 export const isPermitted = (permissions: string[], type: string) => {
