--- app/client/src/pages/organization/invite.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import InviteUsersForm from "./InviteUsersForm";
-
-export const Invite = () => {
-  // const handleInviteUsersSubmit = (values: InviteUsersFormValues) => {};
-  return (
-    <React.Fragment>
-      <h2>Invite Users</h2>
-      <InviteUsersForm />
-    </React.Fragment>
-  );
-};
-
-export default Invite;
