+++ app/client/src/reducers/uiReducers/orgReducer.ts
@@ -72,13 +72,6 @@ const orgReducer = createReducer(initialState, {
       isFetchAllUsers: false,
     },
   }),
-  [ReduxActionTypes.INVITE_USERS_TO_ORG_SUCCESS]: (
-    state: OrgReduxState,
-    action: ReduxAction<OrgUser[]>,
-  ) => ({
-    ...state,
-    orgUsers: [...state.orgUsers, ...action.payload],
-  }),
   [ReduxActionTypes.FETCH_ALL_ROLES_SUCCESS]: (
     state: OrgReduxState,
     action: ReduxAction<Org[]>,
